Fix code after review
The overridden method onOptionsItemSelected() can be safely deleted
because the app does not use it. It takes a lot of space and this
reduction would make the code more compacted and therefore readable. It
is also a distraction from the content that matters and we anticipate
that this method won't be needed during the application development.

Notice, that there is a lag when clicking immediately more than one
button. A user has to wait for the second and subsequent toasts to be
displayed. It is very simple to enhance, the method .cancel() makes a
toast disappear (it's important to check whether the toast is null and
to declare an appropriate member variable).